pre-compile for Windows with node-pre-gyp
A lot of the issues reported are due to Windows users not having
a proper C++/python build environment for native modules. I looked at
a few other popular Node.js modules and it looks like a good solution is
to pre-compile just for Windows using node-pre-gyp. This can be done
easily using AppVeyor.